pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = ["Yauheni_Liubimov <0.5TELOO@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"
django-debug-toolbar = "^3.8.1"
django-taggit = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py310"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.artifacts.*
    | .*\.github.*
    | .*\.idea.*
    | .*\.vscode.*
    | .*docs.*
)
'''


[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".artifacts",
    ".github",
    ".idea",
    ".vscode",
    "docs",
]
